fw-6473, move base tests for controlled site content APIs to own module
