+++ firstvoices/backend/tests/test_apis/test_song_api.py
@@ -5,7 +5,9 @@ import pytest
 from backend.models.constants import Role, Visibility
 from backend.models.song import Lyric, Song
 from backend.tests import factories
-from backend.tests.test_apis.base.base_api_test import BaseControlledSiteContentApiTest
+from backend.tests.test_apis.base.base_controlled_site_api import (
+    BaseControlledSiteContentApiTest,
+)
 from backend.tests.test_apis.base.base_media_test import RelatedMediaTestMixin
